Render new VAFCHart in GenomicOverview
